Allow upload link to work for subdirectory of upload folder.
Currently the behavior is such that the directory specified in the
two requests have to match, which makes the argument for the second
request completely useless. If this were to be the desired behavior
then the `parent_dir` parameter should just be ignored from the second
request completely. Alternatively, this commit allows upload to a subdirectory
of the `p` parameter from the first request again.
This matches the old behavior better without allowing uploading to
an arbitrary directory.

If the current behavior is also desired for some use cases,
it should be added as new request or
with a new parameter on the first request.

Fix regression from #286.

Fix haiwen/seahub#4441