illumos: syscall.Dup3 not implemented in go for solaris
In the go programming language Dup2 and Dup3 are not implemented for solaris
(aka illumos) based operating systems. As long as this is not provided by go
a workaround would be suitable.

As the return of dup3() syscall is not used dup2() could be used. which is
could be workarounded by using the syscall.Syscall implementation.