add config option for multi-instance setups
We've been successfully using Seafile for many years, with multiple instances on one server in parallel, each running on separate fileserver ports, which worked beautifully. With Seafile 12 [1] a hardcoded SeahubURL was introduced which caused us troubles with our multi-instance deployment and might interfere with other services on a host as well. This pull request introduces a new config option in fileserver (**seahub_url**) that allows overriding the hardcoded **127.0.0.1:8000** but does not alter the code path in case the option is not set

[1] https://github.com/haiwen/seafile-server/pull/699/files/678e0dc2f51a1cadb054437760a7387c2a9877ce